+++ changes.md
@@ -4,6 +4,7 @@
 
 * Fix an implementation detail which could throw exceptions over .cljc files.
 * Silence a `println` over a certain code path.
+* Improve `:reflection` linter for an edge case related to .cljc files.
 
 ## Changes from 0.8.1 to 0.9.0